nix-profile.sh: don't subscribe user to channel
There is already code in the Nix installer script to subscribe the user
to a channel, but the code is bypassed when nix-profile.sh (which is
sourced by the installer) has already subscribed the user to a channel.

Another reason for removing this code is that if the user runs
"nix-channel --remove nixpkgs", with the intention to build from source
(or for whatever reason), on the next login the user is again subscribed
to nixpkgs channel. That seems like an anti-feature to me.

Update "tests/nix-profile.sh" to reflect that ~/.nix-channels is no
longer created by this script.